backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Redis / Caching
redis==5.2.1

# Background tasks
celery==5.4.0

# HTTP client (for data ingestion)
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
orjson==3.10.13

# Rate limiting
slowapi==0.1.9

# CSV/data
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
